Refactor plotting into module based on #32
